% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AllPossiblyObsH0_CONT_C}
\alias{AllPossiblyObsH0_CONT_C}
\title{Finite population sample space given an observed dataset.}
\usage{
AllPossiblyObsH0_CONT_C(obs_y0x0z0, obs_y1x0z0, obs_y0x0z1, obs_y1x0z1,
  obs_y0x1z1, obs_y1x1z1)
}
\arguments{
\item{obs_y0x0z0}{Number of observed individuals with Y=0, X=0, Z=0.}

\item{obs_y1x0z0}{Number of observed individuals with Y=1, X=0, Z=0.}

\item{obs_y0x0z1}{Number of observed individuals with Y=0, X=0, Z=1.}

\item{obs_y1x0z1}{Number of observed individuals with Y=1, X=0, Z=1.}

\item{obs_y0x1z1}{Number of observed individuals with Y=0, X=1, Z=1.}

\item{obs_y1x1z1}{Number of observed individuals with Y=1, X=1, Z=1.}
}
\value{
All possibly observable datasets in a list format.
}
\description{
Sample space of all possibly observable datasets given an observed dataset,
    assuming only Compliers and Never Takers in the population.
    This is the C++ function for the R wrapper function
    \code{\link{AllPossiblyObsH0_CONT}}.
}

