% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finitepop.R
\name{expand.grid.DT}
\alias{expand.grid.DT}
\title{Expand.grid using the data.table package.}
\usage{
expand.grid.DT(seq1, seq2, seq3 = NA, seq4 = NA, col.names = NA)
}
\arguments{
\item{seq1}{Vector of values.}

\item{seq2}{Vector of values.}

\item{seq3}{Vector of values.}

\item{seq4}{Vector of values.}

\item{col.names}{Names of columns.}
}
\value{
A data.table with all possible combinations (Cartesian product)
   of the elements in the input vector sequences.
}
\description{
Expand.grid (\code{\link{expand.grid}}) using the data.table package,
   with up to 4 supplied vectors.
}
\examples{
expand.grid.DT(1:10, 100:110)
expand.grid.DT(1:10, 100:110, col.names=c("A", "B"))
expand.grid.DT(1:10, 100:110, 11:13, 1:2)
}

