% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPvalueshypergeoC_allpsi_CONT}
\alias{GetPvalueshypergeoC_allpsi_CONT}
\title{Exact finite population p-values under the sharp null for Compliers.}
\usage{
GetPvalueshypergeoC_allpsi_CONT(n_y0x0z0_H0, n_y1x0z0_H0, n_y0x0z1_H0,
  n_y1x0z1_H0, n_y0x1z1_H0, n_y1x1z1_H0, n_NTy0_H0, n_CONR_H0, n_COAR_H0,
  n_NTy1_H0, critical_regions)
}
\arguments{
\item{n_y0x0z0_H0}{Number of individuals with Y=0, X=0, Z=0.}

\item{n_y1x0z0_H0}{Number of individuals with Y=1, X=0, Z=0.}

\item{n_y0x0z1_H0}{Number of individuals with Y=0, X=0, Z=1.}

\item{n_y1x0z1_H0}{Number of individuals with Y=1, X=0, Z=1.}

\item{n_y0x1z1_H0}{Number of individuals with Y=0, X=1, Z=1.}

\item{n_y1x1z1_H0}{Number of individuals with Y=1, X=1, Z=1.}

\item{n_NTy0_H0}{Total number of Never Takers with Y=0.}

\item{n_CONR_H0}{Total number of Compliers of response type
Never Recover.}

\item{n_COAR_H0}{Total number of Compliers of response type
Always Recover.}

\item{n_NTy1_H0}{Total number of Never Takers with Y=1.}

\item{critical_regions}{Matrix denoting whether each
possibly observed dataset (rows) is in a critical region
(columns).}
}
\value{
A matrix of p-values correspnding to each population
    (row) and the critical region (columns).
}
\description{
Find the exact population-specific p-values
    under the sharp null for Compliers,
    for each compatible population with only Compliers and Never Takers.
    This is the C++ function for the R wrapper function
    \code{\link{Get_pvalues_CONT}}.
}

